+++ docs/src/examples/query/components/query-component/page.js
@@ -1,0 +1,9 @@
+import { $ } from '@semantic-ui/query';
+
+$('.increment').on('click', () => {
+  $('ui-counter').component().increment();
+});
+
+$('.reset').on('click', () => {
+  $('ui-counter').component().reset();
+});