+++ docs/src/examples/query/components/query-setting/component.js
@@ -4,12 +4,12 @@ const css = await getText('./component.css');
 const template = await getText('./component.html');
 
 const defaultSettings = {
-  name: 'World',
-  color: 'blue',
+  message: 'Hello!',
+  active: false,
 };
 
 defineComponent({
-  tagName: 'ui-profile',
+  tagName: 'ui-display',
   template,
   css,
   defaultSettings,