+++ docs/src/examples/query/events/query-focus/component.js
@@ -4,13 +4,13 @@ const css = await getText('./component.css');
 const template = await getText('./component.html');
 
 const defaultSettings = {
-  name: 'World',
-  color: 'blue',
+  label: ''
 };
 
 defineComponent({
-  tagName: 'ui-profile',
+  tagName: 'example-field',
   template,
   css,
   defaultSettings,
+  delegatesFocus: true,
 });